+++ app/src/main/java/com/hongmyeoun/goldcalc/view/setting/settingPage/content/ItemBox.kt
@@ -1,0 +1,76 @@
+package com.hongmyeoun.goldcalc.view.setting.settingPage.content
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.width
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.Icon
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import com.hongmyeoun.goldcalc.ui.theme.LightGrayBG
+import com.hongmyeoun.goldcalc.view.characterDetail.normalTextStyle
+import com.hongmyeoun.goldcalc.view.characterDetail.titleTextStyle
+
+@Composable
+fun ItemBox(
+    title: String,
+    content: @Composable () -> Unit
+) {
+    Text(
+        text = title,
+        style = titleTextStyle(fontSize = 15.sp)
+    )
+    Spacer(modifier = Modifier.height(8.dp))
+
+    Column(
+        modifier = Modifier
+            .fillMaxWidth()
+            .background(
+                color = LightGrayBG,
+                shape = RoundedCornerShape(8.dp)
+            )
+    ) {
+        content()
+    }
+
+    Spacer(modifier = Modifier.height(16.dp))
+}
+
+@Composable
+fun Item(
+    itemTitle: String,
+    icon: Int,
+    onClicked: () -> Unit
+) {
+    Row(
+        modifier = Modifier
+            .fillMaxWidth()
+            .clickable { onClicked() }
+            .padding(top = 12.dp, bottom = 12.dp, start = 16.dp, end = 16.dp),
+        verticalAlignment = Alignment.CenterVertically
+    ) {
+        Icon(
+            painter = painterResource(id = icon),
+            tint = Color.White,
+            contentDescription = "아이콘"
+        )
+        Spacer(modifier = Modifier.width(8.dp))
+
+        Text(
+            text = itemTitle,
+            style = normalTextStyle(fontSize = 15.sp)
+        )
+    }
+}
